requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
apprise==1.9.3
asgi-lifespan==2.1.0
asttokens==3.0.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.2
blinker==1.9.0
cachetools==5.5.1
celery==5.5.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
coolname==2.2.0
cryptography==45.0.2
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.1
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
ecdsa==0.19.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastapi-utils==0.8.0
filetype==1.2.0
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.174.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.2.2
griffe==1.7.3
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.8.2
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kombu==5.5.4
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.74
langchain-experimental==0.3.4
langchain-google-genai==2.0.10
langchain-openai==0.3.3
langchain-postgres==0.0.15
langchain-text-splitters==0.3.8
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.3.45
lxml==5.4.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.4
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
narwhals==1.42.0
nest-asyncio==1.6.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.61.0
openpyxl==3.1.5
opentelemetry-api==1.33.1
orjson==3.10.15
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pgvector==0.3.6
pillow==11.2.1
platformdirs==4.3.6
plotly==6.3.0
prefect==3.4.2
prometheus_client==0.22.0
prompt_toolkit==3.0.50
propcache==0.3.2
proto-plus==1.25.0
protobuf==5.29.3
psutil==5.9.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.10
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.26.5
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
readchar==4.2.1
redis==6.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.25.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
streamlit-pdf-viewer==0.0.26
streamlit-plotly-events==0.0.6
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.3.0
uv==0.7.6
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
